btl_states_menus, fio, DT fixups

+++ src/world/area_end/end_00/parade_setup.c
@@ -71,7 +71,7 @@ API_CALLABLE(N(UpdateCameraScroll)) {
     Camera* camera = &gCameras[gCurrentCameraID];
 
     camera->panActive = TRUE;
-    camera->movePos.x += PARADE_SCROLL_RATE / DT;
+    camera->movePos.x += PARADE_SCROLL_RATE / (f32) DT;
     return ApiStatus_DONE2;
 }
 
@@ -92,7 +92,7 @@ API_CALLABLE(N(AddScrollToNpcPos)) {
         *npc = get_npc_unsafe(evt_get_variable(script, *args++));
     }
 
-    (*npc)->pos.x += PARADE_SCROLL_RATE / DT;
+    (*npc)->pos.x += PARADE_SCROLL_RATE / (f32) DT;
 
     return ApiStatus_BLOCK;
 }
